feat: isolated test cases
