+++ packages/market-matcher/src/test/Matcher.test.ts
@@ -1,379 +1,65 @@
-import { assert } from 'chai';
-import Web3 from 'web3';
-import moment from 'moment';
-import dotenv from 'dotenv';
-
-import { migrateUserRegistryContracts } from '@energyweb/user-registry/contracts';
-import { migrateAssetRegistryContracts } from '@energyweb/asset-registry/contracts';
-import { migrateCertificateRegistryContracts } from '@energyweb/origin/contracts';
-import { migrateMarketRegistryContracts } from '@energyweb/market/contracts';
 import {
+    AssetConsumingRegistryLogic,
     AssetProducingRegistryLogic,
-    ProducingAsset,
-    AssetConsumingRegistryLogic
+    ProducingAsset
 } from '@energyweb/asset-registry';
-import { Agreement, Demand, MarketLogic, Supply } from '@energyweb/market';
+import { Demand, MarketLogic, Supply } from '@energyweb/market';
 import { Certificate, CertificateLogic } from '@energyweb/origin';
-import { buildRights, Role, UserLogic } from '@energyweb/user-registry';
+import { UserLogic } from '@energyweb/user-registry';
 import {
-    Compliance,
     Configuration,
-    Currency,
-    TimeFrame,
-    Unit,
     ContractEventHandler,
-    EventHandlerManager
+    Currency,
+    EventHandlerManager,
+    Unit
 } from '@energyweb/utils-general';
+import { assert } from 'chai';
 
-import { startMatcher, IMatcherConfig } from '..';
-import { logger } from '../Logger';
+import { startMatcher } from '..';
+import {
+    deploy,
+    deployDemand,
+    deployCertificate,
+    deployAsset,
+    deploySupply,
+    deployAgreement
+} from './TestEnvironment';
 
-const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
-
-async function waitForConditionAndAssert(
-    conditionCheckFunction: () => Promise<boolean> | boolean,
-    assertFunction: () => Promise<void> | void,
-    interval: number,
-    timeout: number
-): Promise<void> {
-    let timePassed = 0;
-
-    while (timePassed < timeout) {
-        if (await conditionCheckFunction()) {
-            await assertFunction();
-
-            return;
-        }
-
-        await sleep(interval);
-        timePassed += interval;
-    }
-
-    await assertFunction();
-}
-
-describe('Test StrategyBasedMatcher', async () => {
-    dotenv.config({
-        path: '.env.test'
-    });
-
-    const web3: Web3 = new Web3(process.env.WEB3);
-    const deployKey: string = process.env.DEPLOY_KEY;
-
-    const privateKeyDeployment = deployKey.startsWith('0x') ? deployKey : `0x${deployKey}`;
-    const accountDeployment = web3.eth.accounts.privateKeyToAccount(privateKeyDeployment).address;
-
-    console.log(`acc-deployment: ${accountDeployment}`);
-
-    let smartMeterRead = 0;
-
-    const assetOwnerPK = '0xd9bc30dc17023fbb68fe3002e0ff9107b241544fd6d60863081c55e383f1b5a3';
-    const assetOwnerAddress = web3.eth.accounts.privateKeyToAccount(assetOwnerPK).address;
-
-    const assetSmartMeterPK = '0xc4b87d68ea2b91f9d3de3fcb77c299ad962f006ffb8711900cb93d94afec3dc3';
-    const assetSmartMeter = web3.eth.accounts.privateKeyToAccount(assetSmartMeterPK).address;
-
-    const traderPK = '0xca77c9b06fde68bcbcc09f603c958620613f4be79f3abb4b2032131d0229462e';
-    const accountTrader = web3.eth.accounts.privateKeyToAccount(traderPK).address;
-
-    const issuerPK = '0x622d56ab7f0e75ac133722cc065260a2792bf30ea3265415fe04f3a2dba7e1ac';
-    const issuerAccount = web3.eth.accounts.privateKeyToAccount(issuerPK).address;
-
-    const matcherConfig: IMatcherConfig = {
-        web3Url: process.env.WEB3,
-        offChainDataSourceUrl: process.env.BACKEND_URL,
-        marketContractLookupAddress: '',
-        matcherAccount: {
-            address: accountDeployment,
-            privateKey: privateKeyDeployment
-        }
-    };
-
-    const deployUserRegistry = async () => {
-        const userContracts = await migrateUserRegistryContracts(web3, privateKeyDeployment);
-        const userContractLookupAddress = (userContracts as any).UserContractLookup;
-
-        const userLogic = new UserLogic(web3, (userContracts as any).UserLogic);
-        await userLogic.createUser(
-            'propertiesDocumentHash',
-            'documentDBURL',
-            accountDeployment,
-            'admin',
-            { privateKey: privateKeyDeployment }
-        );
-
-        await userLogic.setRoles(
-            accountDeployment,
-            buildRights([
-                Role.UserAdmin,
-                Role.AssetAdmin,
-                Role.AssetManager,
-                Role.Trader,
-                Role.Matcher
-            ]),
-            { privateKey: privateKeyDeployment }
-        );
-
-        await userLogic.createUser(
-            'propertiesDocumentHash',
-            'documentDBURL',
-            accountTrader,
-            'trader',
-            { privateKey: privateKeyDeployment }
-        );
-
-        await userLogic.setRoles(accountTrader, buildRights([Role.Trader]), {
-            privateKey: privateKeyDeployment
-        });
-
-        await userLogic.createUser(
-            'propertiesDocumentHash',
-            'documentDBURL',
-            assetOwnerAddress,
-            'assetOwner',
-            { privateKey: privateKeyDeployment }
-        );
-        await userLogic.setRoles(assetOwnerAddress, buildRights([Role.AssetManager]), {
-            privateKey: privateKeyDeployment
-        });
-
-        await userLogic.createUser(
-            'propertiesDocumentHash',
-            'documentDBURL',
-            issuerAccount,
-            'issuer',
-            { privateKey: privateKeyDeployment }
-        );
-
-        await userLogic.setRoles(issuerAccount, buildRights([Role.Issuer]), {
-            privateKey: privateKeyDeployment
-        });
-
-        return { userLogic, userContractLookupAddress };
-    };
-
-    const deployAssetRegistry = async (userContractLookupAddress: string) => {
-        const assetRegistryContracts = await migrateAssetRegistryContracts(
-            web3,
-            userContractLookupAddress,
-            privateKeyDeployment
-        );
-        const assetProducingRegistry = new AssetProducingRegistryLogic(
-            web3 as any,
-            (assetRegistryContracts as any).AssetProducingRegistryLogic
-        );
-        const assetContractLookupAddress = (assetRegistryContracts as any).AssetContractLookup;
-
-        return { assetProducingRegistry, assetContractLookupAddress };
-    };
-
-    const deployCertificateRegistry = async (assetContractLookupAddress: string) => {
-        const certificateRegistryContracts = await migrateCertificateRegistryContracts(
-            web3,
-            assetContractLookupAddress,
-            privateKeyDeployment
-        );
-        const certificateLogic = new CertificateLogic(
-            web3 as any,
-            (certificateRegistryContracts as any).CertificateLogic
-        );
-
-        const originContractLookupAddress = (certificateRegistryContracts as any)
-            .OriginContractLookup;
-
-        return { certificateLogic, originContractLookupAddress };
-    };
-
-    const deployMarket = async (
-        assetContractLookupAddress: string,
-        originContractLookupAddress: string,
-        userLogic: UserLogic
-    ) => {
-        const deployedContracts = await migrateMarketRegistryContracts(
-            web3 as any,
-            assetContractLookupAddress,
-            originContractLookupAddress,
-            privateKeyDeployment
-        );
-
-        const marketLogicAddress: string = (deployedContracts as any).MarketLogic;
-
-        const marketLogic = new MarketLogic(web3, marketLogicAddress);
-        const marketContractLookupAddress = (deployedContracts as any).MarketContractLookup;
-
-        await userLogic.createUser(
-            'propertiesDocumentHash',
-            'documentDBURL',
-            marketLogicAddress,
-            'matcher',
-            { privateKey: privateKeyDeployment }
-        );
-
-        await userLogic.setRoles(marketLogicAddress, buildRights([Role.Matcher]), {
-            privateKey: privateKeyDeployment
-        });
-
-        return { marketLogic, marketContractLookupAddress };
-    };
-
-    const deploy = async () => {
-        const { userLogic, userContractLookupAddress } = await deployUserRegistry();
-        const { assetProducingRegistry, assetContractLookupAddress } = await deployAssetRegistry(
-            userContractLookupAddress
-        );
-        const { certificateLogic, originContractLookupAddress } = await deployCertificateRegistry(
-            assetContractLookupAddress
-        );
-
-        const { marketLogic } = await deployMarket(
-            assetContractLookupAddress,
-            originContractLookupAddress,
-            userLogic
-        );
-
-        return {
-            blockchainProperties: {
-                activeUser: {
-                    address: accountTrader,
-                    privateKey: traderPK
-                },
-                userLogicInstance: userLogic,
-                producingAssetLogicInstance: assetProducingRegistry,
-                marketLogicInstance: marketLogic,
-                certificateLogicInstance: certificateLogic,
-                web3
-            },
-            offChainDataSource: {
-                baseUrl: process.env.BACKEND_URL
-            },
-            logger
-        } as Configuration.Entity<
-            MarketLogic,
-            AssetProducingRegistryLogic,
-            AssetConsumingRegistryLogic,
-            CertificateLogic,
-            UserLogic
-        >;
-    };
-
-    const deployDemand = async (config: Configuration.Entity) => {
-        const traderConfig = changeUser(config, {
-            address: accountTrader,
-            privateKey: traderPK
-        });
-
-        const demandOffChainProps: Demand.IDemandOffChainProperties = {
-            timeFrame: TimeFrame.hourly,
-            maxPricePerMwh: 150,
-            currency: Currency.USD,
-            location: ['Thailand;Central;Nakhon Pathom'],
-            assetType: ['Solar'],
-            minCO2Offset: 10,
-            otherGreenAttributes: 'string',
-            typeOfPublicSupport: 'string',
-            energyPerTimeFrame: 1 * Unit.MWh,
-            registryCompliance: Compliance.EEC,
-            startTime: moment().unix(),
-            endTime: moment()
-                .add(1, 'hour')
-                .unix()
-        };
-
-        await Demand.createDemand(demandOffChainProps, traderConfig);
-
-        return Demand.getDemandListLength(traderConfig);
-    };
-
-    const deployAsset = (config: Configuration.Entity) => {
-        
-    }
-
-    const changeUser = (
+describe('Market-matcher e2e tests', async () => {
+    const assertMatched = (
         config: Configuration.Entity,
-        activeUser: { address: string; privateKey: string }
-    ) =>
-        ({
-            ...config,
-            blockchainProperties: { ...config.blockchainProperties, activeUser }
-        } as Configuration.Entity);
+        demand: Demand.Entity,
+        certificate: Certificate.Entity,
+        requiredEnergy: number,
+        done: Mocha.Done
+    ) => {
+        const marketContractEventHandler = new ContractEventHandler(
+            config.blockchainProperties.marketLogicInstance,
+            0
+        );
 
-    describe('Setup', () => {
-        
-        it('should onboard an asset', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountDeployment,
-                privateKey: privateKeyDeployment
-            };
+        marketContractEventHandler.onEvent('DemandPartiallyFilled', async (event: any) => {
+            const { _demandId, _entityId, _amount } = event.returnValues;
 
-            const assetProps: ProducingAsset.IOnChainProperties = {
-                smartMeter: { address: assetSmartMeter },
-                owner: { address: assetOwnerAddress },
-                lastSmartMeterReadWh: 0,
-                active: true,
-                lastSmartMeterReadFileHash: 'lastSmartMeterReadFileHash',
-                propertiesDocumentHash: null,
-                url: null,
-                maxOwnerChanges: 3
-            };
+            if (_demandId === demand.id) {
+                assert.equal(_entityId, certificate.id);
+                assert.equal(_amount, requiredEnergy);
 
-            const assetPropsOffChain: ProducingAsset.IOffChainProperties = {
-                facilityName: 'MatcherTestFacility',
-                operationalSince: 0,
-                capacityWh: 10,
-                country: 'Thailand',
-                address:
-                    '95 Moo 7, Sa Si Mum Sub-district, Kamphaeng Saen District, Nakhon Province 73140',
-                gpsLatitude: '14.059500',
-                gpsLongitude: '99.977800',
-                timezone: 'Asia/Bangkok',
-                assetType: 'Solar',
-                complianceRegistry: Compliance.EEC,
-                otherGreenAttributes: '',
-                typeOfPublicSupport: ''
-            };
+                const updatedCertificate = await certificate.sync();
+                assert.equal(updatedCertificate.owner, demand.demandOwner);
 
-            asset = await ProducingAsset.createAsset(assetProps, assetPropsOffChain, conf);
-            assert.equal(await ProducingAsset.getAssetListLength(conf), 1);
+                done();
+            }
         });
 
-        it('should onboard an supply', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: assetOwnerAddress,
-                privateKey: assetOwnerPK
-            };
-
-            await Supply.createSupply(
-                {
-                    url: null,
-                    propertiesDocumentHash: null,
-                    assetId: '0'
-                },
-                {
-                    price: 150,
-                    currency: Currency.USD,
-                    availableWh: 1e6,
-                    timeFrame: TimeFrame.hourly
-                },
-                conf
-            );
-
-            assert.equal(await Supply.getSupplyListLength(conf), 1);
-        });
-
-        it('no certificate has been created', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountTrader,
-                privateKey: traderPK
-            };
-            assert.equal(await Certificate.getCertificateListLength(conf), 0);
-        });
-
-        it('starts the matcher', async () => {
-            await startMatcher(matcherConfig);
-        });
-    });
+        const eventHandlerManager = new EventHandlerManager(1000, config);
+        eventHandlerManager.registerEventHandler(marketContractEventHandler);
+        eventHandlerManager.start();
+    };
 
     describe('Certificate -> Demand matching tests', () => {
+        const requiredEnergy = 1 * Unit.MWh;
+
         let config: Configuration.Entity<
             MarketLogic,
             AssetProducingRegistryLogic,
@@ -381,268 +67,102 @@ describe('Test StrategyBasedMatcher', async () => {
             CertificateLogic,
             UserLogic
         >;
-        let demandId: string;
+        let demand: Demand.Entity;
+        let asset: ProducingAsset.Entity;
+        let certificate: Certificate.Entity;
 
         before(async () => {
-            config = await deploy();
-            demandId = (await deployDemand(config)).toString();
-            assetId = 
-        });
+            const configuration = await deploy();
+            const { matcherConfig } = configuration;
 
-        it('creates a smart meter reading', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: assetSmartMeter,
-                privateKey: assetSmartMeterPK
-            };
-            const energy = 1 * Unit.MWh;
-            smartMeterRead += energy;
+            config = configuration.config;
+            demand = await deployDemand(config, requiredEnergy);
+            asset = await deployAsset(config);
+            certificate = await deployCertificate(config, asset.id, requiredEnergy);
 
-            const producingAsset = await new ProducingAsset.Entity(asset.id, conf).sync();
-            await producingAsset.saveSmartMeterRead(smartMeterRead, 'newMeterRead');
+            await startMatcher(matcherConfig);
 
-            await certificateLogic.requestCertificates(0, 0, {
-                privateKey: assetOwnerPK
-            });
-
-            await certificateLogic.approveCertificationRequest(0, {
-                privateKey: issuerPK
-            });
-        });
-
-        it('certificate has been created', async () => {
-            assert.equal(await Certificate.getCertificateListLength(conf), 1);
-
-            const certificate = await new Certificate.Entity('0', conf).sync();
-            assert.equal(certificate.owner, assetOwnerAddress);
-        });
-
-        it('certificate has been published for sale', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: assetOwnerAddress,
-                privateKey: assetOwnerPK
-            };
-
-            let certificate = await new Certificate.Entity('0', conf).sync();
-
-            await certificate.publishForSale(1, Currency.USD);
-            certificate = await certificate.sync();
-
-            assert.isTrue(certificate.forSale);
-            assert.equal(await Certificate.getCertificateListLength(conf), 1);
-        });
-
-        it('certificate owner is the trader after successful match', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountDeployment,
-                privateKey: privateKeyDeployment
-            };
-
-            await waitForConditionAndAssert(
-                async () => {
-                    const certificate = await new Certificate.Entity('0', conf).sync();
-
-                    return certificate.owner === accountTrader;
-                },
-                async () => {
-                    const certificate = await new Certificate.Entity('0', conf).sync();
-                    assert.equal(certificate.owner, accountTrader);
-                },
-                1000,
-                30000
+            await certificate.publishForSale(
+                demand.offChainProperties.maxPricePerMwh / 100,
+                demand.offChainProperties.currency
             );
         });
-    });
 
-    describe('Agreement -> Certificate matching tests', () => {
-        it('should create an agreement', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountTrader,
-                privateKey: traderPK
-            };
-
-            const startTime = moment().unix();
-
-            const agreementOffChainProps: Agreement.IAgreementOffChainProperties = {
-                start: startTime,
-                end: startTime + startTime,
-                price: 150,
-                currency: Currency.USD,
-                period: 10,
-                timeFrame: TimeFrame.hourly
-            };
-
-            const agreementProps: Agreement.IAgreementOnChainProperties = {
-                propertiesDocumentHash: null,
-                url: null,
-                demandId: '0',
-                supplyId: '0'
-            };
-
-            await Agreement.createAgreement(agreementProps, agreementOffChainProps, conf);
-
-            conf.blockchainProperties.activeUser = {
-                address: assetOwnerAddress,
-                privateKey: assetOwnerPK
-            };
-
-            const agreement: Agreement.Entity = await new Agreement.Entity('0', conf).sync();
-            await agreement.approveAgreementSupply();
-        });
-
-        it('creates a smart meter reading', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: assetSmartMeter,
-                privateKey: assetSmartMeterPK
-            };
-
-            const energy = 1.5 * Unit.MWh;
-            smartMeterRead += energy;
-
-            console.log(`smartMeterRead=${smartMeterRead}`);
-
-            const producingAsset = await new ProducingAsset.Entity(asset.id, conf).sync();
-            await producingAsset.saveSmartMeterRead(smartMeterRead, 'newMeterRead2');
-
-            await certificateLogic.requestCertificates(0, 1, {
-                privateKey: assetOwnerPK
-            });
-
-            await certificateLogic.approveCertificationRequest(1, {
-                privateKey: issuerPK
-            });
-        });
-
-        it('certificate has been created', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountTrader,
-                privateKey: traderPK
-            };
-            assert.equal(await Certificate.getCertificateListLength(conf), 2);
-        });
-
-        it('certificate is not for sale', async () => {
-            const certificate = await new Certificate.Entity('1', conf).sync();
-
-            assert.isFalse(certificate.forSale);
-        });
-
-        it('a certificate has been split', async () => {
-            await sleep(20000);
-
-            assert.equal(await Certificate.getCertificateListLength(conf), 4);
-        });
-
-        it('asset owner is still the owner of the original certificate', async () => {
-            const certificate = await new Certificate.Entity('1', conf).sync();
-            assert.equal(await certificate.getOwner(), assetOwnerAddress);
-        });
-
-        it('trader is owner of the split certificates', async () => {
-            const certificate1 = await new Certificate.Entity('2', conf).sync();
-            assert.equal(await certificate1.getOwner(), accountTrader);
-
-            const certificate2 = await new Certificate.Entity('3', conf).sync();
-            assert.equal(await certificate2.getOwner(), assetOwnerAddress);
-        });
+        it('certificate should be matched with existing demand', done => {
+            assertMatched(config, demand, certificate, requiredEnergy, done);
+        }).timeout(20000);
     });
 
     describe('Demand -> Certificate matching tests', () => {
-        let certificateId = '';
-        const energy = 1 * Unit.MWh;
+        const requiredEnergy = 1 * Unit.MWh;
+        const price = 150;
+        const currency = Currency.USD;
 
-        it('archives previous demand', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: accountTrader,
-                privateKey: traderPK
-            };
+        let config: Configuration.Entity<
+            MarketLogic,
+            AssetProducingRegistryLogic,
+            AssetConsumingRegistryLogic,
+            CertificateLogic,
+            UserLogic
+        >;
+        let demand: Demand.Entity;
+        let asset: ProducingAsset.Entity;
+        let certificate: Certificate.Entity;
 
-            const demand = new Demand.Entity('0', conf);
-            await demand.changeStatus(Demand.DemandStatus.ARCHIVED);
-        });
+        before(async () => {
+            const configuration = await deploy();
+            const { matcherConfig } = configuration;
 
-        it('creates a smart meter reading', async () => {
-            conf.blockchainProperties.activeUser = {
-                address: assetSmartMeter,
-                privateKey: assetSmartMeterPK
-            };
+            config = configuration.config;
+            asset = await deployAsset(config);
+            certificate = await deployCertificate(config, asset.id, requiredEnergy);
+            await certificate.publishForSale(price / 100, currency);
 
-            smartMeterRead += energy;
-            console.log(`smartMeterRead=${smartMeterRead}`);
+            await startMatcher(matcherConfig);
 
-            const producingAsset = new ProducingAsset.Entity(asset.id, conf);
-            await producingAsset.saveSmartMeterRead(smartMeterRead, 'newMeterRead3');
-
-            const reads = await producingAsset.getSmartMeterReads();
-            const index = reads.length - 1;
-
-            await certificateLogic.requestCertificates(Number(asset.id), index, {
-                privateKey: assetOwnerPK
-            });
-
-            const requestIndex = (await certificateLogic.getCertificationRequests()).length - 1;
-
-            await certificateLogic.approveCertificationRequest(requestIndex, {
-                privateKey: issuerPK
-            });
-
-            conf.blockchainProperties.activeUser = {
-                address: assetOwnerAddress,
-                privateKey: assetOwnerPK
-            };
-
-            certificateId = ((await Certificate.getCertificateListLength(conf)) - 1).toString();
-
-            const certificate = await new Certificate.Entity(certificateId, conf).sync();
-
-            await certificate.publishForSale(1, Currency.USD);
-
-            assert.equal(certificate.energy, energy);
-        });
-
-        it('should create a demand', async () => {
-            await sleep(5000); // wait for matcher to try to match published certificate
-
-            conf.blockchainProperties.activeUser = {
-                address: accountTrader,
-                privateKey: traderPK
-            };
-
-            const demandOffChainProps: Demand.IDemandOffChainProperties = {
-                timeFrame: TimeFrame.hourly,
-                maxPricePerMwh: 150,
-                currency: Currency.USD,
-                location: ['Thailand;Central;Nakhon Pathom'],
-                assetType: ['Solar'],
-                minCO2Offset: 10,
-                otherGreenAttributes: 'string',
-                typeOfPublicSupport: 'string',
-                energyPerTimeFrame: energy,
-                registryCompliance: Compliance.EEC,
-                startTime: moment().unix(),
-                endTime: moment()
-                    .add(1, 'hour')
-                    .unix()
-            };
-
-            await Demand.createDemand(demandOffChainProps, conf);
-            assert.equal(await Demand.getDemandListLength(conf), 2);
+            demand = await deployDemand(config, requiredEnergy, price, currency);
         });
 
         it('demand should be matched with existing certificate', done => {
-            const marketContractEventHandler = new ContractEventHandler(marketLogic, 0);
+            assertMatched(config, demand, certificate, requiredEnergy, done);
+        }).timeout(20000);
+    });
 
-            marketContractEventHandler.onEvent('DemandPartiallyFilled', async (event: any) => {
-                const { _demandId, _entityId, _amount } = event.returnValues;
-                if (_demandId === '1') {
-                    assert.equal(_entityId, certificateId);
-                    assert.equal(_amount, 1000 * Unit.kWh);
-                    done();
-                }
-            });
+    describe('Agreement -> Certificate matching tests', () => {
+        const requiredEnergy = 1 * Unit.MWh;
+        const price = 150;
+        const currency = Currency.USD;
 
-            const eventHandlerManager = new EventHandlerManager(1000, conf);
-            eventHandlerManager.registerEventHandler(marketContractEventHandler);
-            eventHandlerManager.start();
-        }).timeout(10000);
+        let config: Configuration.Entity<
+            MarketLogic,
+            AssetProducingRegistryLogic,
+            AssetConsumingRegistryLogic,
+            CertificateLogic,
+            UserLogic
+        >;
+        let demand: Demand.Entity;
+        let asset: ProducingAsset.Entity;
+        let certificate: Certificate.Entity;
+        let supply: Supply.Entity;
+
+        before(async () => {
+            const configuration = await deploy();
+            const { matcherConfig } = configuration;
+
+            config = configuration.config;
+            asset = await deployAsset(config);
+            certificate = await deployCertificate(config, asset.id, requiredEnergy);
+
+            supply = await deploySupply(config, asset.id, requiredEnergy, price, currency);
+            demand = await deployDemand(config, requiredEnergy, price, currency);
+
+            await deployAgreement(config, demand.id, supply.id, price, currency);
+
+            await startMatcher(matcherConfig);
+        });
+
+        it('demand should be matched with existing certificate', done => {
+            assertMatched(config, demand, certificate, requiredEnergy, done);
+        }).timeout(20000);
     });
 });
